Bring master (by unofficial convention the most recent stable version of Vision) up to at least the same level as its most recent release (staging-8.1.1/release-8.1.1).

+++ software/src/master/src/kernel/VkSocketAddress.cpp
@@ -184,7 +184,7 @@ bool VkSocketAddress::constructGenericAddress (
 	m_iAddress.inOpenVisionFormat.sa_family
     );
     size_t index = 0;
-    u_long nextByte;
+    unsigned int  nextByte;
     int	   nextOffset;
 
     while (
